refactoring- created container for task components outside of App
